pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = {extras = ["standart"], version = "^0.34.0"}
pydantic = {extras = ["email"], version = "^2.11.2"}
pydantic-settings = "^2.8.1"
sqlalchemy = "^2.0.40"
asyncio = "^3.4.3"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
alembic = "^1.15.2"
v = "^1"
passlib = "^1.7.4"
sqladmin = "^0.20.1"
pyjwt = "^2.10.1"
python-jose = "^3.4.0"
